feat: Implement adaptive equalizer for channel compensation
This commit introduces a significant algorithmic enhancement to the ByteWaves modem by implementing an adaptive equalizer to improve performance in noisy and distorting acoustic environments.

Key changes include:

1.  **Adaptive Equalizer**: An `AdaptiveEqualizer` class has been added, utilizing the `padasip` library's LMS (Least Mean Squares) filter. The equalizer is trained on the incoming preamble signal to learn the channel's characteristics and then applies the learned filter to subsequent data chunks to correct for distortion.

2.  **Dependency**: The `padasip` library has been added as a project dependency in `pyproject.toml`.

3.  **Bug Fixes**: This commit also addresses critical issues from a previous code review:
    - The application entry point (`main`, `start_sending`, `start_receiving` functions) has been restored.
    - The Python version requirement in `pyproject.toml` has been corrected to `>=3.8`.
    - A logic flaw in the equalizer training process has been fixed to ensure the training signals are correctly matched.
