+++ bytewaves/__init__.py
@@ -1,0 +1,1 @@
+# This file marks the `bytewaves` directory as a Python package.
